Changed from double to float, also added get_height() function.
No real reason for float vs double, but I keep typing float instinctively so I gave in lol

+++ app/src/main/java/com/example/climbon/Coordinate.java
@@ -1,10 +1,10 @@
 package com.example.climbon;
 
 public class Coordinate {
-    double x;
-    double y;
+    float x;
+    float y;
 
-    Coordinate(double _x, double _y) {
+    Coordinate(float _x, float _y) {
         x = _x;
         y = _y;
     }
